Pull out session logic
